Improve library cleanup during uninstall; Rename default types file

+++ tools/uninstall_node.sh
@@ -47,11 +47,13 @@ sudo -u postgres psql -c "DROP USER \"${POSTGRESQL_ROLE}\";"
 
 [ -f "${BINFILE}" ] && rm -f "${BINFILE}"
 [ -f "${ETCFILE}" ] && rm -f "${ETCFILE}"
-[ -f "${ETCFILE}.sample" ] && rm -f "${ETCFILE}.sample"
 [ -f "${LOGFILE}" ] && rm "${LOGFILE}"
+[ -d "${ETCPATH}/reds" ] && [ -z "`ls -A "${ETCPATH}/reds"`" ] && rm -rf "${ETCPATH}/reds" 
 
-[ $RMLIB = auto ] && [ -d "${ETCPATH}/reds" ] && [ -z "`ls -A "${ETCPATH}/reds"`" ] && RMLIB=true
-if [ $RMLIB = true ]; then
-    [ -d "${ETCPATH}/reds" ] && rm -rf "${ETCPATH}/reds"
-    [ -d "${LIBPATH}/reds" ] && rm -rf "${LIBPATH}/reds"
+# INFO Remove library
+
+if [ $RMLIB = "auto" ]; then
+    RMLIB=false
+    [ -d "${ETCPATH}/reds" ] && [ -z "`ls -A "${ETCPATH}/reds/*_node.json"`" ] && [ -z "`ls -A "${ETCPATH}/reds/*_pod.json"`" ] && RMLIB=true
 fi
+[ $RMLIB = true ] && [ -d "${LIBPATH}/reds" ] && rm -rf "${LIBPATH}/reds"
